# backend/docs/swagger.json
{
    "schemes": [
        "http",
        "https"
    ],
    "swagger": "2.0",
    "info": {
        "description": "Backend API for IELTS Vocabulary Learning Platform",
        "title": "IELTS Learning API",
        "contact": {},
        "version": "v1"
    },
    "basePath": "/api/v1",
    "paths": {
...
        "/topics/{topicId}": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Return topic metadata, parent band level, lesson list, and authenticated user progress.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Topics"
                ],
                "summary": "Get topic detail",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Topic ID",
                        "name": "topicId",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/response.SuccessResponse"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/topic.Response"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/response.ErrorResponse"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/response.ErrorResponse"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/response.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/response.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/vocabularies": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Return paginated vocabulary items with optional search, filters, and authenticated user progress.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Vocabularies"
                ],
                "summary": "List vocabularies",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Search text",
                        "name": "q",
                        "in": "query"
                    },
                    {
                        "enum": [
                            "BEGINNER",
                            "INTERMEDIATE",
                            "ADVANCED"
                        ],
                        "type": "string",
                        "description": "Difficulty",
                        "name": "difficulty",
                        "in": "query"
                    },
                    {
                        "type": "number",
                        "description": "Target IELTS band",
                        "name": "targetBand",
                        "in": "query"
                    },
                    {
                        "enum": [
                            "NEW",
                            "LEARNING",
                            "REVIEW",
                            "MASTERED"
                        ],
                        "type": "string",
                        "description": "User vocabulary status",
                        "name": "status",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "Page number",
                        "name": "page",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "Page size, max 100",
                        "name": "limit",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/response.SuccessResponse"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/vocabulary.ListResponse"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/response.ErrorResponse"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/response.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/response.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/vocabularies/{vocabularyId}": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Return full dictionary-style vocabulary detail and authenticated user progress.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Vocabularies"
                ],
                "summary": "Get vocabulary detail",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Vocabulary ID",
                        "name": "vocabularyId",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/response.SuccessResponse"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/vocabulary.DetailResponse"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/response.ErrorResponse"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/response.ErrorResponse"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/response.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/response.ErrorResponse"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
...
        "models.DifficultyLevel": {
            "type": "string",
            "enum": [
                "BEGINNER",
                "INTERMEDIATE",
                "ADVANCED"
            ],
            "x-enum-varnames": [
                "DifficultyBeginner",
                "DifficultyIntermediate",
                "DifficultyAdvanced"
            ]
        },
...
        "models.LessonStatus": {
            "type": "string",
            "enum": [
                "LOCKED",
                "UNLOCKED",
                "IN_PROGRESS",
                "COMPLETED"
            ],
            "x-enum-varnames": [
                "LessonStatusLocked",
                "LessonStatusUnlocked",
                "LessonStatusInProgress",
                "LessonStatusCompleted"
            ]
        },
...
        "models.VocabularyStatus": {
            "type": "string",
            "enum": [
                "NEW",
                "LEARNING",
                "REVIEW",
                "MASTERED"
            ],
            "x-enum-varnames": [
                "VocabularyStatusNew",
                "VocabularyStatusLearning",
                "VocabularyStatusReview",
                "VocabularyStatusMastered"
            ]
        },
        "response.APIError": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "string"
                },
                "fields": {
                    "type": "object",
                    "additionalProperties": {
                        "type": "string"
                    }
                },
                "message": {
                    "type": "string"
                }
            }
        },
        "response.ErrorResponse": {
            "type": "object",
            "properties": {
                "error": {
                    "$ref": "#/definitions/response.APIError"
                }
            }
        },
        "response.SuccessResponse": {
            "type": "object",
            "properties": {
                "data": {}
            }
        },
...
        "topic.BandLevelResponse": {
            "type": "object",
            "properties": {
                "bandScore": {
                    "type": "number"
                },
                "id": {
                    "type": "integer"
                },
                "title": {
                    "type": "string"
                }
            }
        },
        "topic.LessonResponse": {
            "type": "object",
            "properties": {
                "description": {
                    "type": "string"
                },
                "estimatedMinutes": {
                    "type": "integer"
                },
                "id": {
                    "type": "integer"
                },
                "lockedReason": {
                    "type": "string"
                },
                "progressPercentage": {
                    "type": "integer"
                },
                "slug": {
                    "type": "string"
                },
                "status": {
                    "$ref": "#/definitions/models.LessonStatus"
                },
                "title": {
                    "type": "string"
                },
                "wordCount": {
                    "type": "integer"
                },
                "xpReward": {
                    "type": "integer"
                }
            }
        },
        "topic.Response": {
            "type": "object",
            "properties": {
                "bandLevel": {
                    "$ref": "#/definitions/topic.BandLevelResponse"
                },
                "lessons": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/topic.LessonResponse"
                    }
                },
                "summary": {
                    "$ref": "#/definitions/topic.SummaryResponse"
                },
                "topic": {
                    "$ref": "#/definitions/topic.TopicResponse"
                }
            }
        },
        "topic.SummaryResponse": {
            "type": "object",
            "properties": {
                "completedLessons": {
                    "type": "integer"
                },
                "progressPercentage": {
                    "type": "integer"
                },
                "totalLessons": {
                    "type": "integer"
                },
                "totalXP": {
                    "type": "integer"
                }
            }
        },
        "topic.TopicResponse": {
            "type": "object",
            "properties": {
                "color": {
                    "type": "string"
                },
                "description": {
                    "type": "string"
                },
                "emoji": {
                    "type": "string"
                },
                "icon": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "slug": {
                    "type": "string"
                },
                "title": {
                    "type": "string"
                }
            }
        },
        "vocabulary.BandLevelResponse": {
            "type": "object",
            "properties": {
                "bandScore": {
                    "type": "number"
                },
                "id": {
                    "type": "integer"
                },
                "title": {
                    "type": "string"
                }
            }
        },
        "vocabulary.DetailResponse": {
            "type": "object",
            "properties": {
                "antonyms": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                },
                "audioUrl": {
                    "type": "string"
                },
                "bandLevel": {
                    "$ref": "#/definitions/vocabulary.BandLevelResponse"
                },
                "collocations": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                },
                "difficulty": {
                    "$ref": "#/definitions/models.DifficultyLevel"
                },
                "exampleSentences": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                },
                "frequency": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "ieltsUsage": {
                    "type": "string"
                },
                "ipa": {
                    "type": "string"
                },
                "masteryScore": {
                    "type": "integer"
                },
                "meaningEn": {
                    "type": "string"
                },
                "meaningVi": {
                    "type": "string"
                },
                "partOfSpeech": {
                    "type": "string"
                },
                "primaryMeaning": {
                    "type": "string"
                },
                "progress": {
                    "$ref": "#/definitions/vocabulary.ProgressDetailResponse"
                },
                "rating": {
                    "type": "integer"
                },
                "relatedForms": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/vocabulary.RelatedFormResponse"
                    }
                },
                "secondaryMeaning": {
                    "type": "string"
                },
                "slug": {
                    "type": "string"
                },
                "synonyms": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                },
                "targetBand": {
                    "type": "number"
                },
                "topic": {
                    "$ref": "#/definitions/vocabulary.TopicResponse"
                },
                "userProgress": {
                    "$ref": "#/definitions/vocabulary.ProgressDetailResponse"
                },
                "word": {
                    "type": "string"
                }
            }
        },
        "vocabulary.ListItemResponse": {
            "type": "object",
            "properties": {
                "difficulty": {
                    "$ref": "#/definitions/models.DifficultyLevel"
                },
                "id": {
                    "type": "integer"
                },
                "ipa": {
                    "type": "string"
                },
                "masteryScore": {
                    "type": "integer"
                },
                "meaningEn": {
                    "type": "string"
                },
                "meaningVi": {
                    "type": "string"
                },
                "partOfSpeech": {
                    "type": "string"
                },
                "progress": {
                    "$ref": "#/definitions/vocabulary.ProgressSummaryResponse"
                },
                "shortDefinition": {
                    "type": "string"
                },
                "slug": {
                    "type": "string"
                },
                "status": {
                    "$ref": "#/definitions/models.VocabularyStatus"
                },
                "targetBand": {
                    "type": "number"
                },
                "topic": {
                    "$ref": "#/definitions/vocabulary.TopicResponse"
                },
                "word": {
                    "type": "string"
                }
            }
        },
        "vocabulary.ListResponse": {
            "type": "object",
            "properties": {
                "items": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/vocabulary.ListItemResponse"
                    }
                },
                "pagination": {
                    "$ref": "#/definitions/vocabulary.PaginationResponse"
                }
            }
        },
        "vocabulary.PaginationResponse": {
            "type": "object",
            "properties": {
                "limit": {
                    "type": "integer"
                },
                "page": {
                    "type": "integer"
                },
                "total": {
                    "type": "integer"
                },
                "totalPages": {
                    "type": "integer"
                }
            }
        },
        "vocabulary.ProgressDetailResponse": {
            "type": "object",
            "properties": {
                "correctCount": {
                    "type": "integer"
                },
                "lastReviewedAt": {
                    "type": "string"
                },
                "learnedAt": {
                    "type": "string"
                },
                "masteryScore": {
                    "type": "integer"
                },
                "nextReviewAt": {
                    "type": "string"
                },
                "reviewCount": {
                    "type": "integer"
                },
                "status": {
                    "$ref": "#/definitions/models.VocabularyStatus"
                },
                "wrongCount": {
                    "type": "integer"
                }
            }
        },
        "vocabulary.ProgressSummaryResponse": {
            "type": "object",
            "properties": {
                "masteryScore": {
                    "type": "integer"
                },
                "nextReviewAt": {
                    "type": "string"
                },
                "reviewCount": {
                    "type": "integer"
                },
                "status": {
                    "$ref": "#/definitions/models.VocabularyStatus"
                }
            }
        },
        "vocabulary.RelatedFormResponse": {
            "type": "object",
            "properties": {
                "partOfSpeech": {
                    "type": "string"
                },
                "word": {
                    "type": "string"
                }
            }
        },
        "vocabulary.TopicResponse": {
            "type": "object",
            "properties": {
                "color": {
                    "type": "string"
                },
                "emoji": {
                    "type": "string"
                },
                "icon": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "slug": {
                    "type": "string"
                },
                "title": {
                    "type": "string"
                }
            }
        }
    },
    "securityDefinitions": {
        "BearerAuth": {
            "description": "Type \"Bearer \" followed by a JWT access token.",
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}